Replit/requirements.txt
streamlit==1.31.1
pandas==2.2.0
plotly==5.18.0
psycopg2-binary==2.9.9
openpyxl==3.1.2
numpy==1.26.3
pillow==10.2.0
requests
pandas
pillow
plotly
psycopg2-binary
streamlit
openpyxl
pandas
openai
openai
twilio
plotly
streamlit
yfinance
anthropic
openai
openai
twilio
trafilatura
trafilatura
twilio
trafilatura
openai
openai
openai
twilio
trafilatura
openai
yfinance
yfinance
memory_profiler
openai
flask
matplotlib
openai
flask
openpyxl
python-dateutil
python-chess
